requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
joblib>=1.3.0
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
databases[asyncpg]>=0.7.0
python-dotenv>=1.0.0
cachetools>=5.3.0
httpx>=0.27.0
redis>=5.0.0
torch>=2.2.0
